refactoring of exaamples
