+++ connection_test.go
@@ -6,7 +6,6 @@ package ydb_test
 import (
 	"context"
 	"errors"
-	"fmt"
 	"testing"
 	"time"
 
@@ -14,39 +13,8 @@ import (
 	grpcCodes "google.golang.org/grpc/codes"
 
 	"github.com/ydb-platform/ydb-go-sdk/v3"
-	"github.com/ydb-platform/ydb-go-sdk/v3/balancers"
 	"github.com/ydb-platform/ydb-go-sdk/v3/trace"
 )
-
-func ExampleOpen() {
-	ctx := context.TODO()
-	db, err := ydb.Open(ctx, "grpc://localhost:2135/local")
-	if err != nil {
-		fmt.Printf("connection failed: %v", err)
-	}
-	defer db.Close(ctx) // cleanup resources
-	fmt.Printf("connected to %s, database '%s'", db.Endpoint(), db.Name())
-}
-
-func ExampleOpen_advanced() {
-	ctx := context.TODO()
-	db, err := ydb.Open(
-		ctx,
-		"grpc://localhost:2135/local",
-		ydb.WithAnonymousCredentials(),
-		ydb.WithBalancer(
-			balancers.PreferLocationsWithFallback(
-				balancers.RandomChoice(), "a", "b",
-			),
-		),
-		ydb.WithSessionPoolSizeLimit(100),
-	)
-	if err != nil {
-		fmt.Printf("connection failed: %v", err)
-	}
-	defer db.Close(ctx) // cleanup resources
-	fmt.Printf("connected to %s, database '%s'", db.Endpoint(), db.Name())
-}
 
 func TestZeroDialTimeout(t *testing.T) {
 	ctx, cancel := context.WithTimeout(context.Background(), time.Second*10)
